agent-test-project/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pymysql==1.1.1
pydantic==2.10.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
